AIDS test of US Global Aids Coord
Source:
https://commons.wikimedia.org/wiki/File:Tobias-AIDS-test.jpg

This image is a work of a United States Department of State employee, taken or made as part of that person's official duties. As a work of the U.S. federal government, the image is in the public domain per 17 U.S.C. § 101 and § 105 and the Department Copyright Information.